+++ seqToCanvas_dot.js
@@ -42,7 +42,15 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		L0_75=Math.floor(L0_mid*1.5),
 		/* dotted line path*/
 		dottedLine=()=>{},
+		/* half width of dotted line */
 		dtl50W=Math.floor(dW),
+		/* PI values */
+		PI=Math.PI,PI2=PI/2,PI3=PI/3,PI4=PI/4,
+		PI5=PI/5.0,PI6=PI/6,PI7=PI/7,PI8=PI/8,
+		PI9=PI/9,PI10=PI/10,PI11=PI/11,PI12=PI/12,
+		PI13=PI/13,PI14=PI/14,PI15=PI/15,
+		/* function that returns half width of a dotted line path based on cosine value */
+		cosF=()=>{},
 		/* === methods group: G === */
 		G={
 			0:()=>{},1:()=>{},2:()=>{},3:()=>{},4:()=>{},
@@ -62,9 +70,9 @@ function seqToCanvas_dot(srcCanvasId,seq){
 	//
 	//function that adds dotted line path to the current path
 	dottedLine=(x0,y0,w50)=>{
-		//x0 and y0: relative midpoint coordinates of a dotted line path
-		//w50: half width of a dotted line path in pixels
-		//     definition: w50 = (line length -1)/2
+		// - x0 and y0: relative midpoint coordinates of a dotted line path
+		// - w50: half width of a dotted line path in pixels
+		//     *definition*: w50 = (line length -1)/2
 			/*
 			*	[... -+-+-+-+-+- ...]
 			*	[... 01010101010 ...]
@@ -89,8 +97,16 @@ function seqToCanvas_dot(srcCanvasId,seq){
 	};
 	//-------------------------------------------------------
 	//
+	//function that returns half width of a dotted line path based on cosine value
+	cosF=(w,theta)=>{
+		// - w: width
+		// - theta: angle in radians
+		return (1.0+Math.cos(+theta))*w/2;
+	};
+	//-------------------------------------------------------
+	//
 	G[0]=(e1,e2,e3,e4)=>{
-		//dot
+		//dot: dotted line size is constant
 		//(e1,e2) -> (e3,e4)
 		//
 		let E1=+parseInt(e1,16),
@@ -115,6 +131,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		E1=E2=E3=E4=i=x0=y0=null;
 		N=dx=dy=null;
 	};
+	//-------------------------------------------------------
 	//
 	G[1]=(e1,e2,e3,e4)=>{
 		//dot
@@ -134,7 +151,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#11${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI));
 			i+=1;
 		}
 		//
@@ -161,7 +178,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#22${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI2));
 			i+=1;
 		}
 		//
@@ -188,7 +205,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#33${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI3));
 			i+=1;
 		}
 		//
@@ -215,7 +232,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#44${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI4));
 			i+=1;
 		}
 		//
@@ -223,7 +240,6 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		E1=E2=E3=E4=i=x0=y0=null;
 		N=dx=dy=null;
 	};
-	//-------------------------------------------------------
 	//
 	G[5]=(e1,e2,e3,e4)=>{
 		//dot
@@ -243,7 +259,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#55${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI5));
 			i+=1;
 		}
 		//
@@ -270,7 +286,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#66${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI6));
 			i+=1;
 		}
 		//
@@ -297,7 +313,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#77${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI7));
 			i+=1;
 		}
 		//
@@ -324,7 +340,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#88${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI8));
 			i+=1;
 		}
 		//
@@ -351,7 +367,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#99${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI9));
 			i+=1;
 		}
 		//
@@ -359,7 +375,6 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		E1=E2=E3=E4=i=x0=y0=null;
 		N=dx=dy=null;
 	};
-	//-------------------------------------------------------
 	//
 	G['a']=(e1,e2,e3,e4)=>{
 		//dot
@@ -379,7 +394,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#aa${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI10));
 			i+=1;
 		}
 		//
@@ -406,7 +421,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#bb${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI11));
 			i+=1;
 		}
 		//
@@ -433,7 +448,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#cc${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI12));
 			i+=1;
 		}
 		//
@@ -460,7 +475,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#dd${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI13));
 			i+=1;
 		}
 		//
@@ -487,7 +502,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#ee${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI14));
 			i+=1;
 		}
 		//
@@ -495,7 +510,6 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		E1=E2=E3=E4=i=x0=y0=null;
 		N=dx=dy=null;
 	};
-	//-------------------------------------------------------
 	//
 	G['f']=(e1,e2,e3,e4)=>{
 		//dot
@@ -515,7 +529,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		ctx.fillStyle=`#ff${e1}${e2}${e3}${e4}`;
 		//
 		while(i<N){
-			dottedLine(x0+i*dx/N,y0+i*dy/N,dtl50W);
+			dottedLine(x0+i*dx/N,y0+i*dy/N,cosF(dtl50W,i*PI15));
 			i+=1;
 		}
 		//
@@ -543,7 +557,7 @@ function seqToCanvas_dot(srcCanvasId,seq){
 		L=arr.length
 	}
 	//
-	slf=ctx=W=H=dW=dH=X0=Y0=el0=el1=el2=el3=el4=arr=L=L0=L0_mid=L0_25=L0_75=dottedLine=dtl50W=G=null;
+	slf=ctx=W=H=dW=dH=X0=Y0=el0=el1=el2=el3=el4=arr=L=L0=L0_mid=L0_25=L0_75=dottedLine=dtl50W=PI=PI2=PI3=PI4=PI5=PI6=PI7=PI8=PI9=PI10=PI11=PI12=PI13=PI14=PI15=cosF=G=null;
 };
 //
 //it converts string into canvas image